HTTP checks for MET public web services, plus heartbeat for check_mk server
Summary:
Check_MK local_check for our public web services. The idea is to run a local_check on a box outside our network to verify that the services are available.  Its not meant to be a comprehensive check of the correctness of these services.

Added services and some documentation

Added .arclint. Fixed linting warnings

Added .gitignore

Test Plan: Run script

Reviewers: #it-geo-tf, kimtj

Reviewed By: #it-geo-tf, kimtj

Subscribers: kimtj

Differential Revision: https://phab.met.no/D883